refactor: rename directories - cra-rxjs-styled-components
