+++ cra-rxjs-styled-components/src/components/details-dropdown/DetailsDropdown.tsx
