+++ cra-rxjs-styled-components/src/components/details-dropdown/index.tsx
