+++ cra-rxjs-styled-components/src/components/issue/Issue/Issue.view.tsx
@@ -1,5 +1,5 @@
-import IssueTabHeader from '../IssueTabHeader';
-import IssueCard from '../IssueCard';
+import IssueTabHeader from '../issue-tab-header';
+import IssueCard from '../issue-card';
 import { PaginationContainer, Content, Wrapper } from './Issue.style';
 import type { Issue } from './Issue.type';
 import type { IssueTabValues } from '../types';
