+++ cra-rxjs-styled-components/src/components/issue/issue-card/IssueCard.style.tsx
