+++ cra-rxjs-styled-components/src/components/issue/issue-card/IssueCard.tsx
@@ -1,7 +1,7 @@
 import ClosedIssueIcon from '../../icons/ClosedIssueIcon';
 import MessageIcon from '../../icons/MessageIcon';
 import OpenIssueIcon from '../../icons/OpenIssueIcon';
-import PullRequestIssueInfo from '../../PullRequestIssueInfo';
+import PullRequestIssueInfo from '../../pull-request-issue-info';
 import { IssueCardWrapper } from './IssueCard.style';
 import type { Status } from '../types';
 import { Link } from 'react-router-dom';
