+++ cra-rxjs-styled-components/src/components/issue/issue-card/index.tsx
