+++ cra-rxjs-styled-components/src/components/issue/issue-tab-header/IssueTabHeader.style.tsx
