+++ cra-rxjs-styled-components/src/components/issue/issue-tab-header/index.tsx
