+++ cra-rxjs-styled-components/src/components/misc/privacy-badge/PrivacyBadge.styles.tsx
