+++ cra-rxjs-styled-components/src/components/repo-issues/issue-card-issue-info/index.tsx
