+++ cra-rxjs-styled-components/src/components/repo-issues/issue-card/IssueCard.tsx
@@ -1,7 +1,7 @@
 import ClosedIssueIcon from '../../icons/ClosedIssueIcon';
 import MessageIcon from '../../icons/MessageIcon';
 import OpenIssueIcon from '../../icons/OpenIssueIcon';
-import IssueCardIssueInfo from '../IssueCardIssueInfo';
+import IssueCardIssueInfo from '../issue-card-issue-info';
 import { IssueCardWrapper } from './IssueCard.styles';
 import { Issue } from '../Issues/Issue.type';
 import { State } from '../../../types/types';
