+++ cra-rxjs-styled-components/src/components/repo-issues/issue-card/index.tsx
