+++ cra-rxjs-styled-components/src/helpers/getPullsState.ts
@@ -1,4 +1,4 @@
-import { PullRequest } from '../components/pull-request/PullRequest/PullRequest.type';
+import { PullRequest } from '../components/pull-request/pull-request/PullRequest.type';
 
 export const getPullsState = (pr: PullRequest) => {
   if (pr.state === 'open') {
