+++ cra-rxjs-styled-components/src/routes/repo/repository-pull-request.tsx
@@ -1,4 +1,4 @@
-import PullRequest from '../../components/pull-request/PullRequest';
+import PullRequest from '../../components/pull-request/pull-request';
 
 export default function RepoPullRequest() {
   return <PullRequest />;
